+++ GothicModComposer.UI/Views/GmcSettings.xaml.cs
@@ -1,0 +1,13 @@
+﻿namespace GothicModComposer.UI.Views
+{
+    /// <summary>
+    /// Interaction logic for MainWindow.xaml
+    /// </summary>
+    public partial class GmcSettings
+    {
+        public GmcSettings()
+        {
+            InitializeComponent();
+        }
+    }
+}